restore App.tsx and context destroy by merge / fix some erros
