+++ client/src/components/pageWrapper.tsx
@@ -1,7 +1,7 @@
 import { Container } from "@mui/material";
 
 interface IPageWrapper {
-    title: string,
+    title?: string,
     children: React.ReactNode;
 }
 
@@ -9,7 +9,7 @@ export default function PageWrapper({title, children}: IPageWrapper){
 
     return (
         <Container maxWidth='lg' sx={{display: "flex", flexDirection: "column",justifyContent: "center", alignItems: "center", marginTop: "6rem"}}>
-            <h1 style={{alignSelf: 'start'}}>{title}</h1>
+            {title && <h1 style={{alignSelf: 'start'}}>{title}</h1>}
             {children}
         </Container>
     );
