+++ client/src/pages/Dashboard.tsx
@@ -3,7 +3,6 @@ import PageWrapper from "../components/pageWrapper";
 
 
 export default function Dashboard(){
-
     return (
         <>
             <NavBar />
